added !empty() checks for useragent and ip, otherwise notices are thrown if the user is not logged in.
git-svn-id: https://nom.svn.beanstalkapp.com/microblog/trunk@165 874695d1-85b0-4378-8154-42b2b3dd7dfc

+++ system/application/libraries/Cookie.php
@@ -44,11 +44,11 @@ class Cookie {
 		else 
 		{
 			$data = $this->getAllData();
-			if ($_SERVER['HTTP_USER_AGENT'] != $data['user_agent']) 
+			if (!empty($data['user_agent']) && $_SERVER['HTTP_USER_AGENT'] != $data['user_agent']) 
 			{
 				$this->delete();
 			}
-			if ($_SERVER['REMOTE_ADDR'] != $data['ip']) 
+			if (!empty($data['ip']) && $_SERVER['REMOTE_ADDR'] != $data['ip']) 
 			{
 				$this->delete();
 			}
